+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2024/02/08
+
+- Fixed bug where the annotator was logging duplicate data streams for each video in a session.
+- Fixed bug where the participant status indicators under the session view were not updating properly.
+
 ## [1.0.2] - 2024/02/07
 
 - Fixed bug where the `Coupled` condition for `Video Coupling` would occasionally assign and serve viewers videos for which they were flagged as the `Owner`.